- Nano-ized it into one file - Simplified it - Started on tests
