+++ src/litejavawebsite/Main.java
@@ -2,6 +2,6 @@ package litejavawebsite;
 
 public class Main {
     public static void main(String[] args) {
-        new NanoWebsite().start();
+        new NanoWebsite(9999, "../site").start();
     }
 }
