--- src/litejavawebsite/UriResources.java
@@ -1,6 +1,0 @@
-package litejavawebsite;
-
-public interface UriResources {
-    String getUri();
-    Iterable<String> getResourceNames();
-}
