--- src/site/HelloWorld.html
@@ -1,3 +1,0 @@
-<body>
-    <p id="text">If you see this, then the dev is bad</p>
-</body>
