+++ src/site/HelloWorld.js
@@ -1,1 +1,3 @@
-document.getElementById("text").innerHTML = "You fucking made js work!";
+function onLoad() {
+    document.getElementById("text").innerHTML = "You fucking made js work!";
+}
