+++ src/site/Index.html
@@ -1,0 +1,9 @@
+<!DOCTYPE html>
+<html lang="en">
+    <head>
+        <script type="text/javascript" src="HelloWorld.js"></script>
+    </head>
+    <body onload="onLoad()">
+        <p id="text">If you see this, then the dev is bad</p>
+    </body>
+</html>
