--- src/site/Main.html
@@ -1,67 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <link rel="icon" href="favicon.ico" type="image/x-icon"/>
-        <meta charset="UTF-8">
-        <title>MTG Noodle</title>
-        <script type="text/javascript">
-            function loadPage()
-            {
-                var mimesToElement = { "text/html":"div", "text/js":"js", "text/css":"style" }
-                var resources = [];
-
-                /*for (var i = 0; i < resources.length; i++) {
-                    var request = new XMLHttpRequest();
-                    request.onreadystatechange = function () {
-                        if (request.readyState == 4 && request.status == 200) {
-                            var element = document.createElement(mimesToElement[request.responseType]);
-                            element.type = request.responseType;
-                            element.innerHTML = request.responseText
-                            document.body.appendChild(element);
-                        }
-                    }
-                    request.open("GET", resources[i], true);
-                    request.send(null);
-                }*/
-
-                var request1 = new XMLHttpRequest();
-                request1.onreadystatechange = function () {
-                    if (request1.readyState == 4 && request1.status == 200) {
-                        var element = document.createElement(mimesToElement[request1.responseType]);
-                        element.type = request1.responseType;
-                        element.innerHTML = request1.responseText
-                        document.body.appendChild(element);
-                    }
-                }
-                request1.open("GET", resources[0], true);
-                request1.send(null);
-
-                var request2 = new XMLHttpRequest();
-                request2.onreadystatechange = function () {
-                    if (request2.readyState == 4 && request2.status == 200) {
-                        var element = document.createElement(mimesToElement[request2.responseType]);
-                        element.type = request2.responseType;
-                        element.innerHTML = request2.responseText
-                        document.body.appendChild(element);
-                    }
-                }
-                request2.open("GET", resources[1], true);
-                request2.send(null);
-
-                var request3 = new XMLHttpRequest();
-                request3.onreadystatechange = function () {
-                    if (request3.readyState == 4 && request3.status == 200) {
-                        var element = document.createElement(mimesToElement[request3.responseType]);
-                        element.type = request3.responseType;
-                        element.innerHTML = request3.responseText
-                        document.body.appendChild(element);
-                    }
-                }
-                request3.open("GET", resources[2], true);
-                request3.send(null);
-            }
-        </script>
-    </head>
-    <body onload="loadPage()">
-    </body>
-</html>
