+++ src/site/NotFound.html
@@ -1,3 +1,9 @@
-<body>
-    <img class="background" src="NotFound.png" />
-</body>
+<!DOCTYPE html>
+<html lang="en">
+    <head>
+        <link rel="stylesheet" href="NotFound.css"></link>
+    </head>
+    <body>
+        <img class="background" src="NotFound.png" />
+    </body>
+</html>
